backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
alembic==1.12.1
pydantic==2.5.1
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
openpyxl==3.1.2
xlrd==2.0.1
PyYAML==6.0.1
Jinja2==3.1.2
email-validator==2.1.0.post1
python-dateutil==2.8.2
transformers==4.30.2
torch==2.1.1
spacy==3.7.2
spacy-transformers==1.2.5
sentence-transformers==2.2.1
huggingface-hub==0.16.4
stable-baselines3==2.1.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
cymem==2.0.8
preshed==3.0.9
thinc==8.2.2
wasabi==1.1.2
srsly==2.4.8
catalogue==2.0.10
typer==0.9.0
pathy==0.10.3
smart-open==6.4.0
murmurhash==1.0.10
confection==0.1.4
blis==0.7.11
